editoast: refactor path not found handling
